Add files via upload
AIC - Lower AIC scores are better, and AIC penalizes models that use more parameters. So if two models explain the same amount of variation, the one with fewer parameters will have a lower AIC score and will be the better-fit model.

BIC - Compared to the BIC method (below), the AIC statistic penalizes complex models less. Schwarz's Bayesian Information Criterion (BIC) is a model selection tool. If a model is estimated on a particular data set (training set), BIC score gives an estimate of the model performance on a new, fresh data set (testing set). 

Covariance type is typically nonrobust which means there is no elimination of data to calculate the covariance between features. Covariance shows how two variables move with respect to each other.

Durbin Watson test - One of the assumptions of linear regression is that there is no correlation between the residuals. In other words, the residuals are assumed to be independent. One way to determine if this assumption is met is to perform a Durbin-Watson test, which is used to detect the presence of autocorrelation in the residuals of a regression. The Durbin-Watson statisticlies in the range 0-4. A value of 2 or nearly 2 indicates that there is no first-order autocorrelation. An acceptable range is 1.50 - 2.50. Where successive error differences are small, Durbin-Watson is low (less than 1.50); this indicates the presence of positive autocorrelation.

Indexes - Predict daily price change for SPY by open time. We have 8 predictors and 1 SPY response variable.

US, Asia and EU markets open in the morning, approximately 9 AM - 4 PM. ( US - Nasdaq, SP500, SPY, DJ, Germany - DAX, France - CAC40, Japan - Nikkei, China - Hang Seng, Australia - Aord )

Csv files columns - Open, High, Low, Close, Adj Close, Volume.

Jarque Bera - the Jarque–Bera test is a goodness-of-fit test of whether sample data have the skewness and kurtosis matching a normal distribution.

Kurtosis - Kurtosis is a measure of whether the data are heavy-tailed or light-tailed relative to a normal distribution. That is, data sets with high kurtosis tend to have heavy tails, or outliers. Data sets with low kurtosis tend to have light tails, or lack of outliers. D: Laplace distribution, also known as the double exponential distribution, , excess kurtosis = 3, S: hyperbolic secant distribution, excess kurtosis = 2, L: logistic distribution, excess kurtosis = 1.2, N: normal distribution, excess kurtosis = 0
C: raised cosine distribution, excess kurtosis = −0.593762…, W: Wigner semicircle distribution, excess kurtosis = −1, U: uniform distribution, excess kurtosis = −1.2. If the kurtosis is greater than 3, then the dataset has heavier tails than a normal distribution (more in the tails). If the kurtosis is less than 3, then the dataset has lighter tails than a normal distribution (less in the tails).

Linear Regression Model assumptions - linearity, independence, normality, equal variance. If assumptions are violated you cannot do statistical inference.
Linearity validate -
Independence validate - error is independent. Do Durbin Watson test ( below 1.5 is positive correlation,  between 1.5 and 2.5 is normal).
Normality validate - use QQ plot. Quantile of error distribution and quantile of normal distribution, if normality i not violated there is 45 degrees in plot.
Equal variance validate - cannot be violated to do the do the inference. Plot observed error versus predicted error (RM). 

Maximum drawdown - is a maximum percentage decline in the strategy from the historical peak profit at each point in time. We first compute drawdown, and then the maximum of all drawdowns in the trading period. Maximum drawdown is that risk of mirror for extreme loss of a strategy. To compute a drawdown, we need to compute peak of wealth process. At every time point, we can do it easily using data frame method, cummax. We find maximum drawdown in train and in test are very close. The result shows that if we apply this strategy in test set, the maximum loss from the peak is nine percent. From the mirror of a Sharpe ratio and maximum drawdown, we can tell that the performance of strategy is quite consistent in place of extreme loss. But the return per unit risk is not very consistent.

OLS - In statistics, ordinary least squares (OLS) is a type of linear least squares method for choosing the unknown parameters in a linear regression model (with fixed level-one effects of a linear function of a set of explanatory variables) by the principle of least squares: minimizing the sum of the squares of the differences between the observed dependent variable (values of the variable being observed) in the input dataset and the output of the (linear) function of the independent variable.

Omnibus - Omnibus tests are a kind of statistical test. They test whether the explained variance in a set of data is significantly greater than the unexplained variance, overall. One example is the F-test in the analysis of variance. There can be legitimate significant effects within a model even if the omnibus test is not significant. For instance, in a model with two independent variables, if only one variable exerts a significant effect on the dependent variable and the other does not, then the omnibus test may be non-significant. This fact does not affect the conclusions that may be drawn from the one significant variable. In order to test effects within an omnibus test, researchers often use contrasts.

R2 - R square error - Both RMSE and R2 quantify how well a regression model fits a dataset. The RMSE tells us how well a regression model can predict the value of the response variable in absolute terms while R2 tells us how well a model can predict the value of the response variable in percentage terms.

RMSE - Root mean square error - The RMSE tells us how well a regression model can predict the value of the response variable in absolute terms. RMSE is used more than R Squared. Since the errors are squared before they are averaged, the RMSE gives a relatively high weight to large errors. This means the RMSE is most useful when large errors are particularly undesirable.

Sharpe ratio - Sharpe ratio measures excess return per unit of deviation in an investment asset or trading strategy named after William Sharpe. For example, daily Sharpe ratio is equal to the mean of excess return divided by standard deviation of excess return. Since there are about 252 trading days per year in US stock market, the yearly Sharpe ratio is equal to daily Sharpe ratio multiplied by square root of 252. 

Skew - n probability theory and statistics, skewness is a measure of the asymmetry of the probability distribution of a real-valued random variable about its mean. The skewness value can be positive, zero, negative, or undefined. If skewness is less than -1 or greater than 1, the distribution is highly skewed. If skewness is between -1 and -0.5 or between 0.5 and 1, the distribution is moderately skewed. If skewness is between -0.5 and 0.5, the distribution is approximately symmetric.

SPY - exchange traded fund. SPY is the best-recognized and oldest US listed ETF and typically tops rankings for largest AUM and greatest trading volume. SPY tracks a market cap-weighted index of US large- and mid-cap stocks selected by the S&P Committee.





https://www.statology.org/durbin-watson-test-python/
https://www.coursera.org/learn/python-statistics-financial-analysis/lecture/Joa1c/4-4-multiple-linear-regression-model
https://en.wikipedia.org/wiki/Omnibus_test
https://en.wikipedia.org/wiki/Ordinary_least_squares